improve: upgrade to Gatsby v4
Gatsby 3+ was throwing because we were using Babel packages on client
side that didn't support running in the browser environment. We probably
could have done some more webpack configuration magic to get around
that, but instead we switched to `@babel/standalone`, which is meant to
exactly for these uses cases and is simpler to use.